Run SmartDoor integration tests against live API
